requirements.txt
flask
gunicorn>=20.1.0
nltk
scikit-learn
flask-cors
pymysql
tensorflow
python-dotenv